Flexpert/Flexpert/Flexpert-Design/requirements.txt
Bio
biotite
fair-esm
evaluate
joblib
matplotlib
numpy
omegaconf
packaging
pandas
python_dateutil
pytorch_lightning
PyYAML
requests
safetensors
scikit_learn
scipy
torch
torch_geometric
torcheval
torchmetrics
tqdm
transformers
